requirements.txt
# Django 및 DRF
Django>=4.2.7,<5.0
djangorestframework>=3.14.0
channels>=4.0.0
daphne>=4.0.0

# 데이터베이스
psycopg2-binary>=2.9.9
redis>=5.0.1

# 테스트
pytest>=7.3.1
pytest-django>=4.5.2
pytest-cov>=4.1.0
coverage>=7.2.7

# 문서화
drf-yasg>=1.21.5
Sphinx>=6.2.1

# 데이터 처리
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.1
seaborn>=0.12.2
scipy>=1.11.0
scikit-learn>=1.3.0

# 보안
python-dotenv>=1.0.0
django-cors-headers>=4.2.0

# 모니터링
django-debug-toolbar>=4.1.0
sentry-sdk>=1.28.1

# 배포
gunicorn>=21.2.0
